Add Telegram bot functionality, update requirements and documentation
The changes introduce the main code of the Telegram bot in the new 'main.py' file and the necessary Python packages in 'requirements.txt'. In addition, the README file has been updated with a guide on how to setup and use the bot. Removal of 'test.py', which was not needed, has also been done.